[fix] 모든 주식/코인 정수형으로 수정 [refactor] 전체 코드 포맷 수정 S11P21A108-309

+++ backend/src/main/java/com/example/donzoom/dto/stock/request/StockRequestDto.java
@@ -9,10 +9,10 @@ import lombok.Getter;
 public class StockRequestDto {
 
   private Long stockId;
-  private Float open;
-  private Float close;
-  private Float high;
-  private Float low;
+  private Integer open;
+  private Integer close;
+  private Integer high;
+  private Integer low;
   private LocalDateTime createdAt;
 
 }
