+++ backend/src/main/java/com/example/donzoom/entity/StockWallet.java
@@ -32,17 +32,17 @@ public class StockWallet {
   private Stock stock;
 
   // 총 매입 금액
-  private Float totalInvestedPrice;
+  private Integer totalInvestedPrice;
 
   // 보유한 주식의 수량
   private Integer amount;
 
   // 평균 매입 단가
-  private Float averagePrice;
+  private Integer averagePrice;
 
   @Builder
-  public StockWallet(Wallet wallet, Stock stock, Float totalInvestedPrice, Integer amount,
-      Float averagePrice) {
+  public StockWallet(Wallet wallet, Stock stock, Integer totalInvestedPrice, Integer amount,
+      Integer averagePrice) {
     this.wallet = wallet;
     this.stock = stock;
     this.totalInvestedPrice = totalInvestedPrice;
@@ -51,7 +51,7 @@ public class StockWallet {
   }
 
   // 매수 시 업데이트 로직: 총 매입 금액과 수량, 평단가 업데이트
-  public void updateInvestment(Float totalInvestedPrice, Integer amount) {
+  public void updateInvestment(Integer totalInvestedPrice, Integer amount) {
     this.totalInvestedPrice = totalInvestedPrice;
     this.amount = amount;
     this.averagePrice = totalInvestedPrice / amount;
